Move open/new buttons to overview page (fixes #373)
